Multiple changes on addvoucher.php
It is a waste of resources to calculate the expiration time again and
again for every single voucher when adding multiple vouchers at one
time. The system now calculates it once and uses it for every voucher to
issue.

The user messages where in the wrong place (in the for-section),
creating confusing output